agent: Increase timeout for hotplugged block devices
The timeout of 1 second to wait for a block device to be detected
by the agent was too low in case the system was running on a slow
environment (nested and/or under heavy load).

This commit increase the value from 1 to 10 to leave enough time to
block devices to be detected.

Fixes #218

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>